move example to placeholders
